pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
requests = "^2.23.0"
celery = "^5.2.6"
redis= "^3.5.3"
flower= "^1.0.0"
aiofiles= "^0.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
emails = "^0.5.15"
raven = "^6.10.0"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
alembic = "^1.7.7"
pytest = "^7.1.2"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
python-decouple = "^3.5"
python-dotenv = "^0.20.0"
pandas = "^1.4.0"
xmltodict = "^0.12.0"
zeep = "^4.1.0"
openpyxl = "^3.0.9"
SQLAlchemy = "^1.4.36"
uvicorn = "^0.17.6"
fastapi = "^0.76.0"
pydantic = "^1.9.0"
gunicorn = "^20.1.0"
schedule = "^1.1.0"
httpx = "^0.22.0"
jdatetime = "3.6.4"
celery-jalalicrontab = "^1.0.3"
pyhumps = "^3.5.3"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
# pytest = "^5.4.1"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
black = "^22.3.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
